BAP-21019: Switch to PSR-12 coding standard except explicit constants visibility requirement (#35883)
- used PSR-12 except required visibility for constants
- updated php-cs-fixer config
- added doc block comments for changed classes
- updated php-semver-checker

+++ src/Oro/Bundle/TestFrameworkBundle/Test/WebTestCase.php
@@ -938,7 +938,7 @@ abstract class WebTestCase extends BaseWebTestCase
     public static function generateRandomString($length = 10)
     {
         $random = "";
-        mt_srand((double)microtime() * 1000000);
+        mt_srand((float)microtime() * 1000000);
         $char_list = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
         $char_list .= "abcdefghijklmnopqrstuvwxyz";
         $char_list .= "1234567890_";
